+++ src/main/java/racingcar/cotroller/SetRace.java
@@ -1,16 +1,16 @@
-package racingcar;
+package racingcar.cotroller;
 
-import model.Car;
-import camp.nextstep.edu.missionutils.Console;
+import racingcar.model.Car;
+import racingcar.view.UserInput;
+
 import java.util.ArrayList;
 import java.util.List;
 
 public class SetRace {
     private final List<Car> carsList = new ArrayList<>();
-    public List<Car> createCar() {
-        System.out.println("경주 할 자동차 이름을 입력하세요.(이름은 쉼표(,) 기준으로 구분)");
-        String input = Console.readLine();
 
+    public List<Car> createCar() {
+        String input = UserInput.inputCarName();
         if (input.contains(",")) {
             String[] carNames = input.split(",");
             for (String carName : carNames) {
@@ -25,10 +25,7 @@ public class SetRace {
 
     public int inputMovement() {
         try{
-            System.out.println("이동할 횟수를 입력하세요");
-            String time = Console.readLine();
-            int number = Integer.parseInt(time);
-            return number;
+            return UserInput.inputMovementNumber();
         } catch (NumberFormatException e){
             throw new IllegalArgumentException("숫자를 입력해주세요.");
         }
