+++ src/main/java/racingcar/model/Car.java
@@ -1,4 +1,4 @@
-package model;
+package racingcar.model;
 
 import camp.nextstep.edu.missionutils.Randoms;
